Restructure animation classes
